refactor: normalize kebab-case asset names
