+++ src/index.tsx
@@ -8,7 +8,7 @@ import Store from '@/stores';
 import '@/styles/fonts.css';
 import theme from '@/styles/theme';
 
-import App from './App';
+import App from './app';
 import i18n from './i18n';
 
 const rootElement = document.getElementById('root');
