+++ tests/visual/visual-comparison.authentication.spec.ts-snapshots/uk-large-mobile-2-firefox-linux.png
