+++ tests/visual/visual-comparison.authentication.spec.ts-snapshots/uk-large-mobile-chromium-linux.png
