+++ tests/visual/visual-comparison.authentication.spec.ts-snapshots/uk-large-mobile-webkit-linux.png
