+++ tests/visual/visual-comparison.spec.ts-snapshots/uk-large-mobile-2-chromium-linux.png
