+++ tests/visual/visual-comparison.spec.ts-snapshots/uk-large-mobile-2-firefox-linux.png
