+++ tests/visual/visual-comparison.spec.ts-snapshots/uk-large-mobile-chromium-linux.png
