+++ tests/visual/visual-comparison.spec.ts-snapshots/uk-large-mobile-firefox-linux.png
