+++ tests/visual/visual-comparison.spec.ts-snapshots/uk-large-mobile-webkit-linux.png
